+++ server/src/main/java/com/adobe/testing/s3mock/dto/Mode.java
@@ -20,16 +20,17 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 
 /**
- * <a href="https://docs.aws.amazon.com/AmazonS3/latest/API/API_PutObjectLegalHold.html#API_PutObjectLegalHold_RequestSyntax">API Reference</a>.
+ * <a href="https://docs.aws.amazon.com/AmazonS3/latest/API/API_DefaultRetention.html">API Reference</a>.
  */
-public enum Status {
-  ON("ON"),
-  OFF("OFF");
+public enum Mode {
+
+  GOVERNANCE("GOVERNANCE"),
+  COMPLIANCE("COMPLIANCE");
 
   private final String value;
 
   @JsonCreator
-  Status(String value) {
+  Mode(String value) {
     this.value = value;
   }