+++ server/src/main/java/com/adobe/testing/s3mock/store/BucketStore.java
@@ -16,6 +16,8 @@
 
 package com.adobe.testing.s3mock.store;
 
+import com.adobe.testing.s3mock.dto.ObjectLockConfiguration;
+import com.adobe.testing.s3mock.dto.ObjectLockEnabled;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.File;
 import java.io.IOException;
@@ -59,7 +61,7 @@ public class BucketStore {
     this.retainFilesOnExit = retainFilesOnExit;
     this.s3ObjectDateFormat = s3ObjectDateFormat;
     this.objectMapper = objectMapper;
-    initialBuckets.forEach(bucketName -> this.createBucket(bucketName, null));
+    initialBuckets.forEach(bucketName -> this.createBucket(bucketName, false));
   }
 
   /**
@@ -176,7 +178,7 @@ public class BucketStore {
    * @throws IllegalStateException if the bucket cannot be created or the bucket already exists but
    *        is not a directory.
    */
-  public BucketMetadata createBucket(String bucketName, Boolean objectLockEnabled) {
+  public BucketMetadata createBucket(String bucketName, boolean objectLockEnabled) {
     BucketMetadata bucketMetadata = getBucketMetadata(bucketName);
     if (bucketMetadata != null) {
       throw new IllegalStateException("Bucket already exists.");
@@ -189,7 +191,11 @@ public class BucketStore {
       newBucketMetadata.setName(bucketName);
       newBucketMetadata.setCreationDate(s3ObjectDateFormat.format(LocalDateTime.now()));
       newBucketMetadata.setPath(bucketFolder.toPath());
-      newBucketMetadata.setObjectLockEnabled(objectLockEnabled);
+      if (objectLockEnabled) {
+        newBucketMetadata.setObjectLockConfiguration(
+            new ObjectLockConfiguration(ObjectLockEnabled.ENABLED, null)
+        );
+      }
       writeToDisk(newBucketMetadata);
       return newBucketMetadata;
     }
@@ -209,8 +215,12 @@ public class BucketStore {
   }
 
   public Boolean isObjectLockEnabled(String bucketName) {
-    Boolean objectLockEnabled = getBucketMetadata(bucketName).getObjectLockEnabled();
-    return objectLockEnabled != null ? objectLockEnabled : false;
+    ObjectLockConfiguration objectLockConfiguration =
+        getBucketMetadata(bucketName).getObjectLockConfiguration();
+    if (objectLockConfiguration != null) {
+      return ObjectLockEnabled.ENABLED == objectLockConfiguration.getObjectLockEnabled();
+    }
+    return false;
   }
 
   /**